Fixed ConfigureViper walking
It appears `Flags` returns the full flag set of current command as well as all parents.

When re-binding, it means we rebind a flag multiple time because we get for each child command all the flags that ever been created. To fix this, we now walk using `LocalNonPersistentFlags` so we get flags which are truly local to the is command.

We are going to still get a view of all flags because we recursively walk all commands.